+++ src/view/createEventWithDestination.js
@@ -1,11 +1,27 @@
-export const createEventWithDestination = () => {
+import {getRandomInteger, getRandomArrayElement, getCorrectPreposition} from "../utils.js";
+import {
+  DESCRIPTIONS,
+  EVENT_TYPES,
+  CITIES,
+  ADDITIONAL_OPTIONS,
+  MIN_NUMBER_OPTIONS,
+  MAX_NUMBER_OPTIONS,
+  MIN_NUMBER_PHOTOS,
+  MAX_NUMBER_PHOTOS,
+  MIN_PRICE,
+  MAX_PRICE,
+} from "../const.js";
+
+export const createEventWithDestination = (event) => {
+  const {eventType, destination, destinationInfo, destinationPhotos, additionalOptions, time, price} = event;
+
   return (
     `<form class="trip-events__item  event  event--edit" action="#" method="post">
       <header class="event__header">
         <div class="event__type-wrapper">
           <label class="event__type  event__type-btn" for="event-type-toggle-1">
             <span class="visually-hidden">Choose event type</span>
-            <img class="event__type-icon" width="17" height="17" src="img/icons/flight.png" alt="Event type icon">
+            <img class="event__type-icon" width="17" height="17" src="img/icons/${eventType}.png" alt="Event type icon">
           </label>
           <input class="event__type-toggle  visually-hidden" id="event-type-toggle-1" type="checkbox">
